logout logic

+++ controllers/chatController.js
@@ -4,7 +4,7 @@ const Message = require('../models/messageModel');
 
 async function getMainPage(req, res) {
     const { username } = req.cookies;
-    let chats = await Chat.find({}).sort({ date: -1 });
+    let chats = await Chat.find({userId: username}).sort({ date: -1 });
     let groupedChats = sortChats(chats);
     res.render('main', { chats: groupedChats, username: username});
 }
@@ -18,8 +18,9 @@ async function getChat(req, res) {
 
 async function addChat(req, res) {
     const { title } = req.body;
+    const { username } = req.cookies;
     try {
-        const newChat = await new Chat({ title: title });
+        const newChat = await new Chat({ title: title, userId: username });
         newChat.save();
         res.status(200).json({newChat: newChat})
     } catch (error) {
@@ -49,8 +50,9 @@ async function removeChat(req, res) {
 }
 
 async function removeAllChats(req, res) {
+    const { username } = req.cookies;
     try {
-        await Chat.deleteMany({});
+        await Chat.deleteMany({ userId: username });
         res.sendStatus(200);
     } catch (error) {
         res.sendStatus(500);
